Moved io to drivers and some tidying
